+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.5.0 2026
 
+* [Migrate the kernel plugin sample from MCP tools to Agent capabilities](https://github.com/siyuan-note/siyuan/issues/18638)
+
 ## v0.4.9 2026-8-4
 
 * [Add documentation for the kernels field in plugin.json](https://github.com/siyuan-note/plugin-sample/issues/48)